refactor python to scala functions
