+++ pyzoo/zoo/__init__.py
@@ -18,9 +18,13 @@ from zoo.common.nncontext import *
 from zoo.util.engine import prepare_env
 
 prepare_env()
-JavaCreator.add_creator_class("com.intel.analytics.zoo.pipeline.nnframes.python.PythonNNFrames")
-JavaCreator.add_creator_class("com.intel.analytics.zoo.feature.python.PythonImageFeature")
-JavaCreator.add_creator_class("com.intel.analytics.zoo.pipeline.api.keras.python.PythonAutoGrad")
-JavaCreator.add_creator_class("com.intel.analytics.zoo.models.python.PythonZooModel")
-JavaCreator.add_creator_class("com.intel.analytics.zoo.pipeline.api.keras.python.PythonZooKeras2")
-JavaCreator.add_creator_class("com.intel.analytics.zoo.feature.python.PythonTextFeature")
+with JavaCreator._lock:
+    creator_classes = JavaCreator.get_creator_class()
+    JavaCreator.add_creator_class("com.intel.analytics.zoo.pipeline.nnframes.python.PythonNNFrames")
+    JavaCreator.add_creator_class("com.intel.analytics.zoo.feature.python.PythonImageFeature")
+    JavaCreator.add_creator_class("com.intel.analytics.zoo.pipeline.api.keras.python.PythonAutoGrad")
+    JavaCreator.add_creator_class("com.intel.analytics.zoo.models.python.PythonZooModel")
+    JavaCreator.add_creator_class("com.intel.analytics.zoo.pipeline.api.keras.python.PythonZooKeras2")
+    JavaCreator.add_creator_class("com.intel.analytics.zoo.feature.python.PythonTextFeature")
+    for clz in creator_classes:
+        JavaCreator.add_creator_class(clz)
