+++ zoo/src/main/scala/com/intel/analytics/zoo/common/PythonZoo.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2018 Analytics Zoo Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.intel.analytics.zoo.common
+
+import com.intel.analytics.bigdl.python.api.PythonBigDLKeras
+import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
+
+import scala.reflect.ClassTag
+
+object PythonZoo {
+
+  def ofFloat(): PythonZoo[Float] = new PythonZoo[Float]()
+
+  def ofDouble(): PythonZoo[Double] = new PythonZoo[Double]()
+
+}
+
+
+class PythonZoo[T: ClassTag](implicit ev: TensorNumeric[T]) extends PythonBigDLKeras[T] {
+
+}
